refactoring client side code, adding absolute url paths, including project wide vscode settings, adding eslint and prettier auto save (hopefully that works)
