+++ client/package.json
@@ -38,7 +38,10 @@
   },
   "devDependencies": {
     "autoprefixer": "^10.4.2",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.0.0",
     "postcss": "^8.4.8",
+    "prettier": "^2.6.0",
     "tailwindcss": "^3.0.23"
   }
 }
