+++ client/src/features/exercises/GetSingleExercise.jsx
@@ -1,6 +1,6 @@
-import React, { useState, useEffect } from 'react';
-import { useRouter } from 'next/router';
-import { Button } from  './Button';
+import React, { useState, useEffect } from "react";
+import { useRouter } from "next/router";
+import { Button } from "features/common/Button";
 
 export const GetSingleExercise = () => {
   const [singleExercise, setSingleExercise] = useState(null);
@@ -13,28 +13,29 @@ export const GetSingleExercise = () => {
   useEffect(() => {
     fetch(`http://localhost:4000/exercises/${id}`)
       .then((res) => {
-        if(!res.ok) {
-          throw Error('couldnt fetch exercise')
+        if (!res.ok) {
+          throw Error("couldnt fetch exercise");
         }
         return res.json();
       })
       .then((data) => {
         setSingleExercise(data.exercise);
         setIsLoading(false);
-      }).catch((err) => {
+      })
+      .catch((err) => {
         setError(err.message);
       });
   }, []);
 
   const deleteExercise = async (id) => {
     const response = await fetch(`http://localhost:4000/exercises/${id}`, {
-      method: 'DELETE',
-      headers: { 'Content-Type': 'application/json' },
+      method: "DELETE",
+      headers: { "Content-Type": "application/json" },
     }).then((res) => {
       return res.json();
     });
 
-    if (response.status === 'success') {
+    if (response.status === "success") {
       setSingleExercise(null);
     }
   };
