+++ client/src/features/workouts/CreateWorkout.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import { Input } from './Input';
-import { Select as SelectComponent } from './Select';
-import { Button } from './Button';
+import { Input } from 'features/form/Input';
+import { Select as SelectComponent } from 'features/form/Select';
+import { Button } from 'features/common/Button';
 import Select from 'react-select';
 
 export const CreateWorkout = ({ workouts, setWorkouts, exercises }) => {
