+++ client/src/features/workouts/GetAllWorkouts.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { Button } from  './Button';
+import { Button } from  'features/common/Button';
 import { WorkoutList } from './WorkoutList';
 
 export const GetAllWorkouts = () => {
