+++ client/src/pages/_app.jsx
@@ -1,5 +1,5 @@
-import Layout from '../components/Layout';
-import '../styles/globals.css';
+import Layout from "../features/common/Layout";
+import "../styles/globals.css";
 
 export default function MyApp({ Component, pageProps }) {
   return (
