+++ client/src/pages/index.jsx
@@ -1,15 +1,7 @@
-import React from 'react';
-import { Button } from  '../components/Button';
+import React from "react";
 
 const HomePage = () => {
-  return (
-    <div>
-      HomePage
-      {/* <Button href="/exercises">Exercises</Button>
-        <Button href="/workouts">Workouts</Button>
-        <Button href="/exercises/create">Create exercise</Button> */}
-    </div>
-  );
+  return <div>HomePage</div>;
 };
 
 export default HomePage;
