+++ client/src/pages/workouts/[id].jsx
@@ -1,13 +1,13 @@
-import React from 'react'
-import { GetSingleWorkout } from '../../components/GetSingleWorkout';
+import React from "react";
+import { GetSingleWorkout } from "features/workouts/GetSingleWorkout";
 
 const WorkoutPage = () => {
   return (
     <div>
-        WorkoutPage
-        <GetSingleWorkout />
+      WorkoutPage
+      <GetSingleWorkout />
     </div>
-  )
-}
+  );
+};
 
-export default WorkoutPage
+export default WorkoutPage;
